Separate SQL session management into a dedicated file * Create reuseable MySQL session util * Create reuseable PostgreSQL session util
