+++ common/persistence/sql/sqlplugin/mysql/plugin.go
@@ -25,32 +25,19 @@
 package mysql
 
 import (
-	"strings"
-
-	"github.com/go-sql-driver/mysql"
-	"github.com/iancoleman/strcase"
 	"github.com/jmoiron/sqlx"
 
 	"go.temporal.io/server/common/config"
 	"go.temporal.io/server/common/persistence/sql"
 	"go.temporal.io/server/common/persistence/sql/sqlplugin"
+	"go.temporal.io/server/common/persistence/sql/sqlplugin/mysql/session"
 	"go.temporal.io/server/common/resolver"
 )
 
 const (
 	// PluginName is the name of the plugin
-	PluginName                   = "mysql"
-	isolationLevelAttrName       = "transaction_isolation"
-	isolationLevelAttrNameLegacy = "tx_isolation"
-	defaultIsolationLevel        = "'READ-COMMITTED'"
-	// customTLSName is the name used if a custom tls configuration is created
-	customTLSName = "tls-custom"
+	PluginName = "mysql"
 )
-
-var dsnAttrOverrides = map[string]string{
-	"parseTime":       "true",
-	"clientFoundRows": "true",
-}
 
 type plugin struct{}
 
@@ -92,90 +79,13 @@ func (p *plugin) CreateAdminDB(
 // underlying SQL database. The returned object is to tied to a single
 // SQL database and the object can be used to perform CRUD operations on
 // the tables in the database
-func (p *plugin) createDBConnection(cfg *config.SQL, r resolver.ServiceResolver) (*sqlx.DB, error) {
-	err := registerTLSConfig(cfg)
+func (p *plugin) createDBConnection(
+	cfg *config.SQL,
+	resolver resolver.ServiceResolver,
+) (*sqlx.DB, error) {
+	mysqlSession, err := session.NewSession(cfg, resolver)
 	if err != nil {
 		return nil, err
 	}
-
-	db, err := sqlx.Connect(PluginName, buildDSN(cfg, r))
-	if err != nil {
-		return nil, err
-	}
-	if cfg.MaxConns > 0 {
-		db.SetMaxOpenConns(cfg.MaxConns)
-	}
-	if cfg.MaxIdleConns > 0 {
-		db.SetMaxIdleConns(cfg.MaxIdleConns)
-	}
-	if cfg.MaxConnLifetime > 0 {
-		db.SetConnMaxLifetime(cfg.MaxConnLifetime)
-	}
-
-	// Maps struct names in CamelCase to snake without need for db struct tags.
-	db.MapperFunc(strcase.ToSnake)
-	return db, nil
-}
-
-func buildDSN(cfg *config.SQL, r resolver.ServiceResolver) string {
-	mysqlConfig := mysql.NewConfig()
-
-	mysqlConfig.User = cfg.User
-	mysqlConfig.Passwd = cfg.Password
-	mysqlConfig.Addr = r.Resolve(cfg.ConnectAddr)[0]
-	mysqlConfig.DBName = cfg.DatabaseName
-	mysqlConfig.Net = cfg.ConnectProtocol
-	mysqlConfig.Params = buildDSNAttrs(cfg)
-
-	// https://github.com/go-sql-driver/mysql/blob/v1.5.0/dsn.go#L104-L106
-	// https://github.com/go-sql-driver/mysql/blob/v1.5.0/dsn.go#L182-L189
-	if mysqlConfig.Net == "" {
-		mysqlConfig.Net = "tcp"
-	}
-
-	// https://github.com/go-sql-driver/mysql#rejectreadonly
-	// https://github.com/temporalio/temporal/issues/1703
-	mysqlConfig.RejectReadOnly = true
-
-	return mysqlConfig.FormatDSN()
-}
-
-func buildDSNAttrs(cfg *config.SQL) map[string]string {
-	attrs := make(map[string]string, len(dsnAttrOverrides)+len(cfg.ConnectAttributes)+1)
-	for k, v := range cfg.ConnectAttributes {
-		k1, v1 := sanitizeAttr(k, v)
-		attrs[k1] = v1
-	}
-
-	// only override isolation level if not specified
-	if !hasAttr(attrs, isolationLevelAttrName) &&
-		!hasAttr(attrs, isolationLevelAttrNameLegacy) {
-		attrs[isolationLevelAttrName] = defaultIsolationLevel
-	}
-
-	// these attrs are always overriden
-	for k, v := range dsnAttrOverrides {
-		attrs[k] = v
-	}
-
-	return attrs
-}
-
-func hasAttr(attrs map[string]string, key string) bool {
-	_, ok := attrs[key]
-	return ok
-}
-
-func sanitizeAttr(inkey string, invalue string) (string, string) {
-	key := strings.ToLower(strings.TrimSpace(inkey))
-	value := strings.ToLower(strings.TrimSpace(invalue))
-	switch key {
-	case isolationLevelAttrName, isolationLevelAttrNameLegacy:
-		if value[0] != '\'' { // mysql sys variable values must be enclosed in single quotes
-			value = "'" + value + "'"
-		}
-		return key, value
-	default:
-		return inkey, invalue
-	}
+	return mysqlSession.DB, nil
 }
